sports-alerts/requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
